### 2021-03-16 #3107 @csoneson
👋 @DidierMurilloF - thanks for submitting to JOSS. I will handle your submission, and start by finding a couple of suitable reviewers. If you have any suggestions (e.g., from the list linked in the first post above), please mention them here (without tagging them with `@`).

I noticed that your paper is currently missing a 'Statement of Need' section (see [here](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)). Could you please add that (I think you can break out a part of the 'Summary' section)? Also, please fix the invalid DOI mentioned by Whedon above (remove the `https://doi.org/` prefix). Thanks!
