+++ joss/paper.bib
@@ -32,9 +32,8 @@ Treatment Sets},
     pages = {461-480},
     year={2011},
     keywords = {error degrees of freedom, expected genetic gain, fixed and random effects, optimal design, shrinkage},
-    doi = {https://doi.org/10.1111/j.1467-842X.2011.00642.x},
-    url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1467-842X.2011.00642.x},
-    eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1467-842X.2011.00642.x}
+    doi = {10.1111/j.1467-842X.2011.00642.x},
+    url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1467-842X.2011.00642.x}
 }
 
 @book{Federer_cit,
