+++ joss/paper.md
@@ -33,7 +33,9 @@ bibliography: paper.bib
 
 # Summary
 
-`FielDHub` is an R Shiny design of experiments (DOE) app that aids in the creation of traditional, unreplicated, augmented and partially-replicated [@Cullis2006_cit] designs  applied to agriculture, plant breeding, forestry, animal and biological sciences. One of the problems that life scientists often face is the lack of freely available and user-friendly interactive tools to create designs that fit their needs. A few open-source DOE R packages options exist including agricolae [@agricolae_cit] and blocksdesign [@blocksdesign_cit], but they require users to be familiar with the R programming language and do not have a graphical user interface (GUI). 
+`FielDHub` is an R Shiny design of experiments (DOE) app that aids in the creation of traditional, unreplicated, augmented and partially-replicated [@Cullis2006_cit] designs  applied to agriculture, plant breeding, forestry, animal and biological sciences. One of the problems that life scientists often face is the lack of freely available and user-friendly interactive tools to create designs that fit their needs. A few open-source DOE R packages options exist including agricolae [@agricolae_cit] and blocksdesign [@blocksdesign_cit], but they require users to be familiar with the R programming language and do not have a graphical user interface (GUI).
+
+# Statement of need
 
 `FielDHub` allows users to perform randomizations of field, laboratory, and greenhouse experiments, while providing output via interactive field layouts and tables that can be extracted and saved. This app has a novel design that offers DOE options and features that are not currently available in most software applications.  Users are guided in each step of the DOE platform in an interactive interface, which includes a feature that helps to generate randomizations with an option to simulate data for a response variable. This last feature makes it suitable for teaching and evaluation purposes, where instructors can use the graphical dynamic user interface and/or use the functions included in the R package for teaching R scripting courses. This app also provides a graphical workflow to import treatment lists and export field books. For field experiments with a strict spatial arrangement, it allows users to specify the dimensions of the field as row and columns, while controlling the percentage of check plots, and obtaining field maps and field books that can be used directly as templates and input files for centralized databases.
 
