### 2021-03-17 #3107 @csoneson
👋 @Prof-ThiagoOliveira, @dlebauer - would you be interested in reviewing this submission for JOSS?

> FielDHub: An R Shiny Package for Design of Experiments in Life Sciences
