### 2021-03-17 #3107 @Prof-ThiagoOliveira
Dear Madam/Sir,

Yes, I would appreciate to review this paper.

Best regards,
Dr. Thiago de Paula Oliveira
Blog: https://prof-thiagooliveira.netlify.com/
​
Researcher in Statistics
The Roslin Institute
University of Edinburgh
