### 2021-03-18 #3122 @whedon
[REVIEW]: FielDHub: An R Shiny Package for Design of Experiments in Life Sciences
**Submitting author:** @DidierMurilloF (<a href="http://orcid.org/0000-0003-3685-4959">Didier Murillo Florez</a>)
**Repository:** <a href="https://github.com/DidierMurilloF/FielDHub" target ="_blank">https://github.com/DidierMurilloF/FielDHub</a>
**Version:** v0.1.0
