### 2021-03-18 #3122 @csoneson
👋 @DidierMurilloF, @Prof-ThiagoOliveira, @dlebauer - this is the review thread for the submission. All of our communications will happen here from now on.

Both reviewers have checklists at the top of this thread. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.
