### 2021-03-19 e5f832d
I fixed some bugs at R version 3.6.0
