+++ R/fct_RCDB_augmented.R
@@ -27,7 +27,8 @@
 #' Macmillan Company.
 #' 
 #' @examples
-#' # Example 1: Generates an ARCBD with 6 blocks, 3 checks for each, and 50 treatments in two locations.
+#' # Example 1: Generates an ARCBD with 6 blocks, 3 checks for each, and 50 treatments 
+#' # in two locations.
 #' ARCBD1 <- RCBD_augmented(lines = 50, checks = 3, b = 6, l = 2, 
 #'                          planter = "cartesian", 
 #'                          plotNumber = c(1,1001),
@@ -39,7 +40,8 @@
 #' ARCBD1$plotNumber
 #' head(ARCBD1$fieldbook, 12)
 #'                    
-#' # Example 2: Generates an ARCBD with 17 blocks, 4 checks for each, and 350 treatments in 3 locations.
+#' # Example 2: Generates an ARCBD with 17 blocks, 4 checks for each, and 350 treatments 
+#' # in 3 locations.
 #' # In this case, we show how to use the option data.
 #' checks <- 4;
 #' list_checks <- paste("CH", 1:checks, sep = "")
