+++ README.md
@@ -1,10 +1,10 @@
-## FielDHub: An R Shiny Package for Design of Experiments in Life Sciences.
+## FielDHub: A Shiny App for Design of Experiments in Life Sciences
 
 
-[![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)
+[![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html)
 
 
-FielDHub is an R Shiny design of experiments (DOE) app that aids in the creation of traditional, unreplicated, augmented and partially-replicated designs applied to agriculture, plant breeding, forestry, animal and biological sciences. 
+A shiny design of experiments (DOE) app that aids in the creation of traditional, unreplicated, augmented and partially-replicated designs applied to agriculture, plant breeding, forestry, animal and biological sciences. 
 
 ## Installation
 
@@ -12,13 +12,6 @@ You can install the dev version of FielDHub from:
 
 ``` r
 devtools::install_github("DidierMurilloF/FielDHub")
-```
-
-or 
-
-
-``` r
-remotes::install_github("DidierMurilloF/FielDHub")
 ```
 
 ## Example
