### 2021-03-19 5e35eb5
fct_CRD-print_summary
