+++ man/print.summary.CRD.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/fct_CRD.R
+\name{print.summary.CRD}
+\alias{print.summary.CRD}
+\title{Print the summary of an \code{CRD} object}
+\usage{
+\method{print}{summary.CRD}(x, ...)
+}
+\arguments{
+\item{x}{an object inheriting from class \code{\link[FielDHub]{CRD}}}
+
+\item{...}{Unused, for extensibility}
+}
+\value{
+an object inheriting from class \code{CRD}
+}
+\description{
+Print summary information on the design parameters, and
+  data frame structure
+}
+\author{
+Thiago de Paula Oliveira,
+  \email{thiago.paula.oliveira@alumni.usp.br}
+}
