+++ man/print.CRD.Rd
@@ -1,0 +1,35 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/fct_CRD.R
+\name{print.CRD}
+\alias{print.CRD}
+\title{Print an \code{CRD} object}
+\usage{
+\method{print}{CRD}(x, n, ...)
+}
+\arguments{
+\item{x}{an object inheriting from class}
+
+\item{n}{a single integer...
+\code{\link[FielDHub]{CRD}}}
+
+\item{...}{further arguments passed to \code{\link{head}}.}
+}
+\value{
+an object inheriting from class \code{CRD}
+}
+\description{
+Prints information about randomization of Completely
+  Randomized Design (CDR)
+}
+\examples{
+# Example 1: Generates a CRD design with 5 treatments and 5 reps each.
+crd1 <- CRD(t = 5, reps = 5, plotNumber = 101,
+seed = 1985, locationName = "Fargo")
+crd1$infoDesign
+print(crd1)
+
+}
+\author{
+Thiago de Paula Oliveira,
+  \email{thiago.paula.oliveira@alumni.usp.br}
+}
