### 2021-03-20 #3122 @Prof-ThiagoOliveira
PR submitted with changes on fct_full_factorial.R. [DidierMurilloF/FielDHub#6](https://github.com/DidierMurilloF/FielDHub/pull/6)
