### 2021-03-20 DidierMurilloF/FielDHub#7 @Prof-ThiagoOliveira
Authorship by function
Could you provide authorship for each function in the package rather than just in the description file? This is an important step to see how authors/contributors have contributed to this package's development. Please, remember that each contributor should appear in the description file. Sorry to be a little bit pedantic about it, but you will see this step is important over the next releases (years). When you describe contributors and authors only on the description file, I assumed they have contributed to all functions, which, probably, is not true.

Please, check the box when you have attributed authorship for the following functions:

- [ ] `get_golem_config()`
- [x] `app_server()`
- [x] `app_ui() `
- [x] `alpha_lattice()`
- [x] `CDR()`
- [x] `diagonal_arrangement()`
- [x] `full_factorial()`
- [x] `incomplete_blocks()`
- [x] `latin_square()`
- [x] `optimized_arrangement()`
- [x] `partially_replicated()`
- [x] `RCBD()`
- [x] `RCBD_augmented()`
- [x] `rectangular_lattice()`
- [x] `row_column()`
- [x] `split_families()`
- [x] `split_plot()`
- [x] `split_split_plot()`
- [x] `square_lattice()`
- [x] `strip_plot()`
