### 2021-03-20 9a7cef2
Summary, print, class and Set names for latin-square function and Class for CRD
