+++ R/fct_latin_square.R
@@ -1,6 +1,6 @@
-#' Generates a Latin Square Design 
+#' Generates a Latin Square Design
 #'
-#' Randomly generates a latin square design of up 10 treatments. 
+#' Randomly generates a latin square design of up 10 treatments.
 #'
 #' @param t Number of treatments.
 #' @param reps Number of full resolvable squares. By default \code{reps = 1}.
@@ -9,12 +9,13 @@
 #' @param seed (optional) Real number that specifies the starting seed to obtain reproducible designs.
 #' @param locationNames (optional) Name for the location.
 #' @param data (optional) Data frame with label list of treatments.
-#' 
+#'
 #' @importFrom stats runif na.omit
 #'
-#' @return A list with information on the design parameters. 
+#' @return A list with information on the design parameters.
 #' @return Data frame with the latin square field book.
-#' 
+#'
+#' @importFrom stats setNames
 #'
 #' @references
 #' Federer, W. T. (1955). Experimental Design. Theory and Application. New York, USA. The
@@ -22,34 +23,34 @@
 #'
 #' @examples
 #' # Example 1: Generates a latin square design with 4 treatments and 2 reps.
-#' latinSq1 <- latin_square(t = 4, 
-#'                          reps = 2, 
-#'                          plotNumber = 101, 
+#' latinSq1 <- latin_square(t = 4,
+#'                          reps = 2,
+#'                          plotNumber = 101,
 #'                          planter = "cartesian",
-#'                          seed = 1980) 
-#' latinSq1$squares
-#' latinSq1$plotSquares
+#'                          seed = 1980)
+#' print(latinSq1)
+#' summary(latinSq1)
 #' head(latinSq1$fieldBook)
-#' 
+#'
 #' # Example 2: Generates a latin square design with 5 treatments and 3 reps.
 #' latin_data <- data.frame(list(ROW = paste("Period", 1:5, sep = ""),
 #'                               COLUMN = paste("Cow", 1:5, sep = ""),
 #'                               TREATMENT = paste("Diet", 1:5, sep = "")))
 #' print(latin_data)
-#' latinSq2 <- latin_square(t = NULL, 
-#'                          reps = 3, 
-#'                          plotNumber = 101, 
+#' latinSq2 <- latin_square(t = NULL,
+#'                          reps = 3,
+#'                          plotNumber = 101,
 #'                          planter = "cartesian",
-#'                          seed = 1981, 
-#'                          data = latin_data) 
+#'                          seed = 1981,
+#'                          data = latin_data)
 #' latinSq2$squares
 #' latinSq2$plotSquares
 #' head(latinSq2$fieldBook)
-#' 
+#'
 #' @export
-latin_square <- function(t = NULL, reps = 1, plotNumber = 101,  planter = "serpentine", 
+latin_square <- function(t = NULL, reps = 1, plotNumber = 101,  planter = "serpentine",
                          seed = NULL, locationNames = NULL, data = NULL) {
-  
+
   if (is.null(seed) || !is.numeric(seed)) seed <- runif(1, min = -50000, max = 50000)
   set.seed(seed)
   if (all(c("serpentine", "cartesian") != planter)) {
@@ -102,12 +103,13 @@ latin_square <- function(t = NULL, reps = 1, plotNumber = 101,  planter = "serpe
   step.random <- vector(mode = "list", length = reps)
   lsd.reps <- vector(mode = "list", length = reps)
   out.ls <- vector(mode = "list", length = reps)
-  plotSquares <- vector(mode = "list", length = reps)
+  plotSquares <- setNames(vector(mode = "list", length = reps),
[…]
