+++ R/fct_incomplete_blocks.R
@@ -1,17 +1,17 @@
-#' Generates a Resolvable Incomplete Block Design 
+#' Generates a Resolvable Incomplete Block Design
 #'
-#' @description Randomly generates a resolvable incomplete block design (IBD) of characteristics (t, k, r). 
+#' @description Randomly generates a resolvable incomplete block design (IBD) of characteristics (t, k, r).
 #' The randomization can be done across locations.
 #'
 #' @param t Number of  treatments.
 #' @param r Number of full blocks (or resolvable replicates) (also number of replicates per treatment).
-#' @param k Size of incomplete blocks (number of units per incomplete block). 
+#' @param k Size of incomplete blocks (number of units per incomplete block).
 #' @param l Number of locations. By default \code{l = 1}.
 #' @param plotNumber Numeric vector with the starting plot number for each location. By default \code{plotNumber = 101}.
 #' @param seed (optional) Real number that specifies the starting seed to obtain reproducible designs.
 #' @param locationNames (optional) Names for each location.
 #' @param data (optional) Data frame with label list of treatments.
-#' 
+#'
 #' @importFrom stats runif na.omit
 #'
 #' @return A list with information on the design parameters.
@@ -25,26 +25,26 @@
 #' @examples
 #' # Example 1: Generates a resolvable IBD of characteristics (t,k,r) = (12,4,2).
 #' # 1-resolvable IBDs
-#' ibd1 <- incomplete_blocks(t = 12, 
-#'                           k = 4, 
-#'                           r = 2, 
+#' ibd1 <- incomplete_blocks(t = 12,
+#'                           k = 4,
+#'                           r = 2,
 #'                           seed = 1984)
 #' ibd1$infoDesign
 #' head(ibd1$fieldBook)
-#' 
+#'
 #' # Example 2: Generates a balanced resolvable IBD of characteristics (t,k,r) = (15,3,7).
 #' # In this case, we show how to use the option data.
 #' treatments <- paste("TX-", 1:15, sep = "")
 #' treatment_list <- data.frame(list(TREATMENT = treatments))
-#' head(treatment_list) 
+#' head(treatment_list)
 #' ibd2 <- incomplete_blocks(t = 15,
-#'                           k = 3, 
-#'                           r = 7, 
+#'                           k = 3,
+#'                           r = 7,
 #'                           seed = 1985,
-#'                           data = treatment_list) 
+#'                           data = treatment_list)
 #' ibd2$infoDesign
-#' head(ibd2$fieldBook)  
-#'              
+#' head(ibd2$fieldBook)
+#'
 #' @export
 incomplete_blocks <- function(t = NULL, k = NULL, r = NULL, l = 1, plotNumber = 101, locationNames = NULL,
                               seed = NULL, data = NULL) {
@@ -65,7 +65,7 @@ incomplete_blocks <- function(t = NULL, k = NULL, r = NULL, l = 1, plotNumber = 
       if (length(t) == 1) {
         if (t == 1 || t < 1) {
           shiny::validate('incomplete_blocks() requires more than one treatment.')
-        } 
+        }
         nt <- t
       }else if ((length(t) > 1)) {
         nt <- length(t)
@@ -74,7 +74,7 @@ incomplete_blocks <- function(t = NULL, k = NULL, r = NULL, l = 1, plotNumber = 
     }else if (is.character(t) || is.factor(t)) {
       if (length(t) == 1) {
[…]

[…]
