### 2021-03-21 DidierMurilloF/FielDHub#8 @Prof-ThiagoOliveira
Output treatment for factorial designs
Dear authors,

Please, include a new argument like the `factorLabels` I have suggested here [(DidierMurilloF/FielDHub#7)](https://github.com/DidierMurilloF/FielDHub/pull/6).  This argument can help end-users to understand the treatment output better.

A suggestion of alternative expected output:

- [x] Function `full_factorial()` - Done!
```
     ID LOCATION PLOT REP FACTOR_A FACTOR_B FACTOR_C   TRT_COMB
1     1   Loc1    101   1       a0       b0       c1   a0*b0*c1
2     2   Loc1    102   1       a1       b0       c1   a1*b0*c1
3     3   Loc1    103   1       a1       b1       c0   a1*b1*c0
```

- [x] Function `split_plot()`:
```
     ID LOCATION    PLOT    REP   WHOLE-PLOT    SUB-PLOT     TRT_COMB
1     1    A        101      1       Fung4       Beans5     Fung4|Beans5
2     2    A        101      1       Fung4       Beans1     Fung4|Beans1
3     3    A        101      1       Fung4       Beans2     Fung4|Beans2
```

- [x] Function `split_split_plot()`:
```
   ID LOCATION PLOT REP WHOLE-PLOT SUB-PLOT SUB-SUB-PLOT TRT_COMB
1   1        A  101   1     IRR_NO    Fung3       Beans3    IRR_NO|Fung3|Beans3
2   2        A  101   1     IRR_NO    Fung3       Beans9    IRR_NO|Fung3|Beans9
3   3        A  101   1     IRR_NO    Fung3      Beans10    IRR_NO|Fung3|Beans10
```

- [x] Function `strip_plot ()` - Done!
