### 2021-03-22 DidierMurilloF/FielDHub#9 @Prof-ThiagoOliveira
Checks argument
Dear authors,

I know we can use checks/tester to determine the effectiveness of manipulation in an experimental design related to social/medial science. Furthermore, checks can also be used in agriculture (mainly in genetics) and generally are commercial varieties because you know their characteristics and associated parameters.

You defined checks as an integer number of genotypes checks or a numeric vector. This argument definition is a bit confusing to me.  Firstly, what do you mean by a numeric vector? What happens when end-users include a numeric vector there? I suggest you re-write this argument explaining to the end-user all the details, not just part of them.

Another point is in the Shiny App. Why did you decide to work with percentage for checks? Why you restricted this percentage to vary from 6.67 to 10%? It is not clear what is the difference between %Checks and Input # of checks. Alternatively, could you describe each shiny input in the dashboard?
