### 2021-03-22 DidierMurilloF/FielDHub#7 @Prof-ThiagoOliveira
Please, check the box when you have attributed authorship for the following functions:

- [ ] `get_golem_config()`
- [ ] ` `
- [ ] ` `
- [ ] ` `
- [ ] ` `
- [ ] ` `
