### 2021-03-22 DidierMurilloF/FielDHub#7 @Prof-ThiagoOliveira
Please, check the box when you have attributed authorship for the following functions:

- [ ] `get_golem_config()`
- [ ] `app_server()`
- [ ] `app_ui() `
- [ ] `alpha_lattice()`
- [ ] `CDR()`
- [ ] `diagonal_arrangement()`
- [ ] `full_factorial()`
- [ ] `incomplete_blocks()`
- [ ] `latin_square()`
- [ ] `optimized_arrangement()`
- [ ] `partially_replicated()`
- [ ] `RCBD()`
- [ ] `RCBD_augmented()`
- [ ] `rectangular_lattice()`
- [ ] `row_column()`
- [ ] `split_families()`
- [ ] `split_plot()`
- [ ] `split_split_plot()`
- [ ] `square_lattice()`
- [ ] `strip_plot()`
