### 2021-03-22 DidierMurilloF/FielDHub#10 @Prof-ThiagoOliveira
Shiny table output
Dear authors,

I realized you have several wide tables as output in your shiny app. When a table is too wide, you can fix certain columns using the `FixedColumns` extension to be visible when you scroll the table horizontally. Thus, I suggest you fix the first column related to row labels (see image) and also include a scroll on the x-axis to facilitate end-user interaction with the output. You can include a scrolling row by adding in the datatable options `scrollX = TRUE`.

![table_design](https://user-images.githubusercontent.com/34395021/112033897-cc2d9900-8b35-11eb-9145-f54e8ca47ef7.png)

Ex.:
```
datatable(
  your_data,
  extensions = 'FixedColumns',
  options = list(
    dom = 't',
    scrollX = TRUE,
    fixedColumns = TRUE
  )
)
```
Does this suggestion make sense to you?
Best regards,
