### 2021-03-22 DidierMurilloF/FielDHub#7 @tatirri
Dear Professor Thiago Oliveira,
Thank you very much for reaching out to us regarding the authorship of these functions. The two main authors who wrote all these functions are Didier Murillo-Florez and Salvador Alejandro Gezan. We will take into account your recommendations and put their names in the formulas above. The other people included in the paper have contributed with ideas, patches, shiny development coding and field design theory and background related to agriculture/biology applications of the design of experiments.

Thanks for your time and all your contributions to this application and to the paper.

Have a wonderful rest of your day,

Ana Maria Heilman-Morales
