### 2021-03-22 DidierMurilloF/FielDHub#10 @tatirri
Professor Thiago Oliveira,

We will start making the changes accordingly to the recommendations you provided. Thanks for your input. AM
