### 2021-03-22 DidierMurilloF/FielDHub#7 @Prof-ThiagoOliveira
Dear Ana Maria,

Thank you for your reply. Please note that people who contributed with some idea to the code should also be included as contributors to that function. So, you can distinguish them by attributing to each person a label such as `[aut]` for author or `[ctb]` for a contributor.

Have a  wonderful rest of your day too!
