### 2021-03-22 DidierMurilloF/FielDHub#11 @Prof-ThiagoOliveira
Shiny development
In the summary of your manuscript, you describe: "One of the problems that life scientists often face is the lack of freely available and user-friendly interactive tools to create designs that fit their needs". I absolutely agree with the author's phrase. Thus, I proposed here a discussion about this topic.

By user-friendly definition, I understand that common attributes of it should be
 1. **Interface should be simple/straightforward**. It should provide quick access to common features or commands.
 2. **Clean**. A good user interface is well-organized (users cannot feel confounded among inputs)
 3. **Intuitive**. The interface must make sense to the average user and require a minimal explanation for using it.
 4. **Reliable**.  It is something that doesn't cause undue frustration for the user.

Let us discuss each of these items now.
  1. The shiny interface that the authors provided looks nice, and the inclusion of a box with all inputs is well defined. It is quite easy for the end-user to interact with the dashboard, and it puts the app as a simple/straightforward tool to create designs.
  2. Although the authors proposed some complex experimental designs, their app seems very clean and with an organized interface.
  3. Unfortunately, some inputs of experimental designs proposed in the app are not intuitive and require some improvement to the end-user to understand the whole idea. I can use myself as an example here. I have never worked with the unreplicated design before be presented to this manuscript/package. In this way, I can tell the authors that a student or people with no background won't use this interface because it is too hard to set up each input means. This is due to the app didn't provide any description of each input. Consequently, the user needs to kind of guess the meaning of each one. To turn the app intuitive, I suggest that the authors work a little more in the description step and perhaps include a concise introduction to each design (say one paragraph).
  4. As the app is sometimes not intuitive, it also causes frustration for the user. However, it is reliable because it does not malfunction or crash.

With these improvements, the application will greatly impact universities and be something easy to use.
