### 2021-03-22 DidierMurilloF/FielDHub#9 @DidierMurilloF
Dear Thiago Oliveira,

We will provide a better explanation of your questions here. First, the reason why we consider the number of checks as an integer is to define the number of genotypes that are being used in the experiment (i.e. 4 genotypes of the total will be checks) and also the Entry numbers for each. For example, if `checks = 4` (integer number), we have four checks, and the function `diagonal_arrangement()` is going to assign `1 2 3 4` as entry numbers for each. In another case, users might want to customize entries for their checks just like `checks = c(297,298,299,300)`, which implies there are 4 checks but now the Entry numbers for each are `297 298 299 300`. In that case, the input was a numeric vector.

In our app, the first four entry numbers will be represented by check or control genotypes. Regarding the percentage of checks and the reason behind why this percentage is restricted is because of the dimension the user will use in the field. We want to provide flexibility to the user to define how many columns and rows they can put in the field and the percentage of check plots in the field will be based on the total amount of entries that are being tested and the number of checks the user is utilizing. Also, breeders usually allocate amounts of plot checks equivalent to `8%, 10%`, or `12 %` of the total available plots. That is why we fixed the percentage of checks approximately between `6%` and `20%`.

I hope this can answer your questions.

Best regards,

Didier murillo
