### 2021-03-22 DidierMurilloF/FielDHub#9 @DidierMurilloF
I know that non-replicated designs can be complex and not very common. But, the truth is that they are very used and useful when you want to test a large number of varieties with restricted quantities of seed, or physical resources. Please, let me share with you a key article in this area.
Clarke, G. P. Y., & Stefanova, K. T. (2011). Optimal design for early-generation plant breeding trials with unreplicated or partially replicated test lines. Australian & New Zealand Journal of Statistics,
53(4), 461–480.

https://onlinelibrary.wiley.com/doi/full/10.1111/j.1467-842X.2011.00642.x

Thanks,

Didier Murillo
