### 2021-03-22 DidierMurilloF/FielDHub#9 @Prof-ThiagoOliveira
@DidierMurilloF Thank you for your explanation. This now makes sense to me. Perhaps, reformulate the description of this argument. What do you think?

I'll read this reference. I agree with you about the complexities of those experiments, which is exactly why the idea of this package is so important.
Best wishes,
