### 2021-03-22 DidierMurilloF/FielDHub#9 @DidierMurilloF
@Prof-ThiagoOliveira  Thank you for seeing the value of this work, especially since there are no packages in R to design these types of experiments.

I'm going to rephrase the description of the input `checks`. I think I should include more details. Thank you!
