### 2021-03-22 DidierMurilloF/FielDHub#7 @DidierMurilloF
Dear Thiago,

I would like to ask you if this is the right way to implement your request. The following information will be in the  R functions description.

#' @author Didier Murillo [aut],
#'         Salvador Gezan [aut],
#'         Ana Heilman [ctb],
#'         Thomas Walk [ctb],
#'         Johan Aparicio [ctb],
#'         Richard Horsley [ctb]

Thank you!
