### 2021-03-23 DidierMurilloF/FielDHub#7 @Prof-ThiagoOliveira
Dear @DidierMurilloF,

Of course, you can differentiate types of contribution by function; however, it can be too informative. My suggestion here is to include names and email only. These are the most common ways I have been seen in R:

1. Example 1:
#' @author Didier Murillo, \email{your email}, Salvador Gezan, Ana Heilman, ...

2. Example 2:
#' @author Didier Murillo, \email{your email}, Salvador Gezan \email{email}, Ana Heilman \email{email}, ...

Perhaps, my suggestion above may have confounded you. Sorry about that.
