### 2021-03-26 DidierMurilloF/FielDHub#11 @tatirri
Dear Thiago,

Thank you for the careful review and relevant suggestions. We have put a lot of effort into these areas and appreciate the feedback. We are actively working to build more interactive guides for more complex experimental designs. We estimate that these changes will take weeks to properly design and incorporate, and we are planning to include them in the next release. As users are actively planning their experiments for 2021, we feel that FielDHub is sufficient for meeting users needs for this upcoming growing season. Delaying release will prevent some of our users from designing their experiments in FielDHub this year. While we remain focused on improving usability and will add more intuitive guides, we believe that it is helpful for users this year and we are seeking to distribute FielDHub as soon as possible. The less complex designs, such as RCBD, are the most widely used, and FielDHub is simple for those use cases. Nevertheless, we are willing to remove references to an intuitive design until we update the guides for the most complex designs if that helps us to expedite publication.

Didier, Ana & Tom
