### 2021-03-26 DidierMurilloF/FielDHub#11 @Prof-ThiagoOliveira
Dear Didier, Ana & Tom,

I absolutely agree with your point of view. There is no reason to barrier the publication due to small (but important) details for the first release. I agree that this suggestion will take some weeks to be included, and I am more than happy to accept the paper when other suggestions I have made were implemented in the package. We can leave this suggestion for the next release. I kindly ask you to make the next realise as soon as possible.

Is this reasonable for you?
Best wishes,

Thiago
