+++ man/alpha_lattice.Rd
@@ -71,3 +71,11 @@ head(alphalattice2$fieldBook, 10)
 Edmondson., R. N. (2021). blocksdesign: Nested and crossed block designs for factorial and
 unstructured treatment sets. https://CRAN.R-project.org/package=blocksdesign
 }
+\author{
+Didier Murillo [aut],
+        Salvador Gezan [aut],
+        Ana Heilman [ctb],
+        Thomas Walk [ctb], 
+        Johan Aparicio [ctb], 
+        Richard Horsley [ctb]
+}
