+++ NAMESPACE
@@ -45,6 +45,7 @@ importFrom(stats,qnorm)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
 importFrom(stats,sd)
+importFrom(stats,setNames)
 importFrom(utils,capture.output)
 importFrom(utils,head)
 importFrom(utils,read.csv)
