+++ R/fct_RCDB_augmented.R
@@ -12,6 +12,7 @@
 #' @param exptName (optional) Name of experiment.
 #' @param locationNames (optional) Name for each location.
 #' @param repsExpt (optional) Number of reps of experiment. By default \code{repsExpt = 1}.
+#' @param random Logical value to randomize treatments or not. By default \code{random = TRUE}.
 #' @param data (optional) Data frame with the labels of treatments.
 #' 
 #' @importFrom stats runif na.omit
@@ -63,7 +64,7 @@
 #' @export
 RCBD_augmented <- function(lines = NULL, checks = NULL, b = NULL, l = 1, planter = "serpentine", 
                            plotNumber = 101, exptName  = NULL, seed = NULL, locationNames = NULL,
-                           repsExpt = 1, data = NULL) {
+                           repsExpt = 1, random = TRUE, data = NULL) {
 
   if (all(c("serpentine", "cartesian") != planter)) {
     stop("Input planter choice is unknown. Please, choose one: 'serpentine' or 'cartesian'.")
@@ -85,7 +86,7 @@ RCBD_augmented <- function(lines = NULL, checks = NULL, b = NULL, l = 1, planter
       shiny::validate("RCBD_augmented() requires input plotNumber to be possitive integers and sorted.")
     }
   }
-
+  
   if (!is.null(data)) {
     data <- as.data.frame(data)
     if (ncol(data) < 2) base::stop("Data input needs at least two columns with: ENTRY and NAME.")
@@ -126,84 +127,126 @@ RCBD_augmented <- function(lines = NULL, checks = NULL, b = NULL, l = 1, planter
     for (expts in expt) {
       Blocks <- vector(mode = "list", length = b)
       layout <- base::matrix(data = 0, nrow = b, ncol = plots_per_block, byrow = TRUE)
-      if (Fillers > 0) {
-        if (Fillers >= (plots_per_block - checks - 2)) {
-          shiny::validate("Number of Filler overcome the amount allowed per block. Please, choose another quantity of blocks.")
-        } 
-        len_cuts <- rep(lines_per_plot, times = b - 1)
-        len_cuts <- c(len_cuts, lines - sum(len_cuts))
-        entries <- as.vector(data[(checks + 1):nrow(data),1])
-        entries <- sample(entries)
-        rand_len_cuts <- sample(len_cuts)
-        lines_blocks <- split_vectors(x = entries, len_cuts = rand_len_cuts)
-        if (b %% 2 == 0) { 
-          if(planter == "serpentine") {
-            layout[1,1:Fillers] <- "Filler"
-          }else{ 
-            layout[1,((ncol(layout) + 1) - Fillers):ncol(layout)] <- "Filler" 
+      if (random) {
+        if (Fillers > 0) {
+          if (Fillers >= (plots_per_block - checks - 2)) {
+            shiny::validate("Number of Filler overcome the amount allowed per block. Please, choose another quantity of blocks.")
           } 
-        }else { 
-          layout[1,((ncol(layout) + 1) - Fillers):ncol(layout)] <- "Filler" 
-        }
-        z <- b
-        w <- 1
-        for (i in b:1) {
-          datos <- sample(c(rep(0, len_cuts[w]), 1:checks))
-          Bi <- sample(datos)
-          ci <- as.vector(layout[i,])
-          ci[ci == 0] <- Bi
-          layout[i,] <- ci
-          Blocks[[z]] <- ci
-          z <- z - 1
-          w <- w + 1
-        }
-        col_checks <- ifelse(layout != 0, 1, 0)
-        v <- which.min(rand_len_cuts)
[…]
