### 2021-04-07 #3122 @dlebauer
@csoneson this is my first review for JOSS. Just learning how this works. One of the first things I usually do when reviewing an article is see if it meets journal requirements. In this case I started by submitting a pr to meet the journal requirements of having an OSI approved license. However it hasn't been merged or addressed (eg with a different license). While this doesn't prevent a full review, I was hoping this would be a first step and would establish the pattern for using issues and prs as a mechanism for review as suggested by JOSS, and which I think is a great idea.
