### 2021-04-07 #3122 @csoneson
Thanks @dlebauer. From the DESCRIPTION file it looks like the package is indeed MIT licensed. @DidierMurilloF - could you clarify the situation here (and add the license to the repository as well)?
