### 2021-04-07 DidierMurilloF/FielDHub#12 @dlebauer
Add link to repository in JOSS paper and contributing information
Per requirements, need to add link to repository in the references section, or add a section such as

```
## Availability and Community Guidelines

The software is available at the Comprehensive R Archive Network CRAN and also at the
GitHub repository https://github.com/DidierMurilloF/FielDHub/. The GitHub repository also contains the source code for this paper.

Users and contributors are welcome to contribute, request features, and report bugs in this GitHub repository.
```

It would also be helpful to have a CONTRIBUTING.md or similar that could be referenced here, and would meet the JOSS requirements https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines. Mozilla provides some ideas https://mozillascience.github.io/working-open-workshop/contributing/
