### 2021-04-07 b243585
Update fct_full_factorial.R
- updating docs to indicate that only one object is returned from function
