### 2021-04-07 DidierMurilloF/FielDHub#14 @dlebauer
Suggestions to improve documentation
Based on JOSS guidelines https://joss.readthedocs.io/en/latest/review_criteria.html#documentation

- it is not clear where to find documentation
  - as an R user I can find the `help(package = 'FielDHub')` and I see this is great!
  - for the community the README can be substantially improved
    - a lot of the content can come directly from the JOSS paper - including the figures and some additional examples ... or have a section on getting help and running examples using `help()` and `example()` functions
    - link to the pkgdown site
 - use `return` only once per function
 - flesh out `run_app` documentation. This is the key user facing function! What are some options that can be passed as `...`?
 - in both the paper and the docs, it could be more clear that there are a large number of useful standalone functions
