### 2021-04-08 b7d8885
Outputs for replications and authorship by functions.
