### 2021-04-08 DidierMurilloF/FielDHub#7 @DidierMurilloF
Dear Profesor Thiago,

I added the information for the authorship by function. Please, could you check it? Thank you!

Didier Murillo.
