### 2021-04-08 DidierMurilloF/FielDHub#14 @Prof-ThiagoOliveira
@dlebauer fair point. @DidierMurilloF, here is a list of all functions I saw more than one `return`.

- [ ] fct_RCDB_augmented.R
- [ ] fct_CRD.R
- [ ] fct_RCBD.R
- [ ] fct_alpha_lattice.R
- [ ] fct_diagonal_arrangement.R
- [ ] fct_full_factorial.R
- [ ] fct_incomplete_blocks.R
- [ ] fct_latin_square.R
- [ ] fct_optimized_arrangement.R
- [ ] fct_full_factorial.R
- [ ] fct_incomplete_blocks.R
- [ ] fct_latin_square.R
- [ ] fct_optimized_arrangement.R
- [ ] fct_partially_replicated.R
- [ ] fct_rectangular_lattice.R
- [ ] fct_row_column.R
- [ ] fct_split_families.R
- [ ] fct_split_plot.R
- [ ] fct_split_split_plot.R
- [ ] fct_square_lattice.R
- [ ] fct_strip_plot.R
