### 2021-04-12 1e469f8
Output treatment for factorial and other designs.
