### 2021-04-12 DidierMurilloF/FielDHub#8 @DidierMurilloF
Dear Thiago,

We included the  `factorLabels` argument into `split_plot()`, `split_split_plot()` and `strip_plot ()` functions. Please, let us know if you have any question. Thank you!

Best,

Didier Murillo
