### 2021-04-12 286173b
Shiny table output (JOSS)
