+++ R/mod_Diagonal.R
@@ -476,9 +476,16 @@ mod_Diagonal_server <- function(id) {
                      'greenyellow', 'blueviolet','deepskyblue','gold','blue', 'red')
         s <- unlist(rand_lines()$Entries)
         rownames(df) <- nrow(df):1
-        options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE,
-                                  scrollX = TRUE, scrollY = "1000px"))
-        DT::datatable(df) %>% 
+        # options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE,
+        #                           scrollX = TRUE, scrollY = "1000px"))
+        options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE, scrollY = "700px"))
+        DT::datatable(df,
+                      extensions = 'FixedColumns',
+                      options = list(
+                        dom = 't',
+                        scrollX = TRUE,
+                        fixedColumns = TRUE
+                      )) %>% 
           DT::formatStyle(paste0(rep('V', ncol(df)), 1:ncol(df)),
                           backgroundColor = DT::styleEqual(c(checks), 
                                                            colores[1:len_checks])) %>% 
