### 2021-04-12 7990fe3
Shiny table output request (JOSS)
