+++ R/mod_Diagonal.R
@@ -333,12 +333,20 @@ mod_Diagonal_server <- function(id) {
       
       df <- as.data.frame(w_map)
       rownames(df) <- nrow(df):1
-      options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE,
-                                scrollX = TRUE, scrollY = "1000px"))
-      DT::datatable(df) %>% 
-        DT::formatStyle(paste0(rep('V', ncol(df)), 1:ncol(df)),
-                        backgroundColor = DT::styleEqual(c(checks,0,"Filler"), 
-                                                         c(colores[1:len_checks],'yellow', 'snow')))
+      # options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE,
+      #                           scrollX = TRUE, scrollY = "1000px"))
+      # DT::datatable(df) %>%
+      options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE, scrollY = "700px"))
+      DT::datatable(df,
+                    extensions = 'FixedColumns',
+                    options = list(
+                      dom = 't',
+                      scrollX = TRUE,
+                      fixedColumns = TRUE
+                    )) %>% 
+      DT::formatStyle(paste0(rep('V', ncol(df)), 1:ncol(df)),
+                      backgroundColor = DT::styleEqual(c(checks,0,"Filler"), 
+                                                       c(colores[1:len_checks],'yellow', 'snow')))
     })
     
     output$dt8 <- DT::renderDT({
@@ -520,19 +528,35 @@ mod_Diagonal_server <- function(id) {
           Visual_ch <- as.numeric(input$Visual_ch)
           df <- as.data.frame(r_map)
           rownames(df) <- nrow(df):1
-          options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE,
-                                    scrollX = TRUE, scrollY = "1000px"))
-          DT::datatable(df) %>%
-            DT::formatStyle(paste0(rep('V', ncol(df)), 1:ncol(df)),
-                            backgroundColor = DT::styleEqual(c(a, Visual_ch), c(my_colors, 'red')))
+          # options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE,
+          #                           scrollX = TRUE, scrollY = "1000px"))
+          # DT::datatable(df) %>%
+          options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE, scrollY = "700px"))
+          DT::datatable(df,
+                        extensions = 'FixedColumns',
+                        options = list(
+                          dom = 't',
+                          scrollX = TRUE,
+                          fixedColumns = TRUE
+                        )) %>% 
+          DT::formatStyle(paste0(rep('V', ncol(df)), 1:ncol(df)),
+                          backgroundColor = DT::styleEqual(c(a, Visual_ch), c(my_colors, 'red')))
         }else{
           df <- as.data.frame(r_map)
           rownames(df) <- nrow(df):1
-          options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE,
-                                    scrollX = TRUE, scrollY = "1000px"))
-          DT::datatable(df) %>%
-            DT::formatStyle(paste0(rep('V', ncol(df)), 1:ncol(df)),
-                            backgroundColor = DT::styleEqual(a, my_colors))
+          # options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE,
+          #                           scrollX = TRUE, scrollY = "1000px"))
+          # DT::datatable(df) %>%
+          options(DT.options = list(pageLength = nrow(df), autoWidth = FALSE, scrollY = "700px"))
+          DT::datatable(df,
+                        extensions = 'FixedColumns',
+                        options = list(
+                          dom = 't',
+                          scrollX = TRUE,
[…]
