### 2021-04-12 DidierMurilloF/FielDHub#10 @DidierMurilloF
Dear Thiago,

I got your suggestions and I added the features to make the table outputs more comprehensible for the end-users.  I already did the push and you can see the changes. Please, let us know if have any question! Thank You!

Best,

Didier Murillo
