### 2021-04-27 DidierMurilloF/FielDHub#12 @DidierMurilloF
Dear @dlebauer,

You mean we should add a new section in the paper just like

## Availability and Community Guidelines

The software is available at the Comprehensive R Archive Network CRAN and also at the
GitHub repository https://github.com/DidierMurilloF/FielDHub/. The GitHub repository also contains the source code for this paper.

Users and contributors are welcome to contribute, request features, and report bugs in this GitHub repository.

We can do that. We just need to confirm that is your request. Thanks

Didier Murillo.
