### 2021-04-27 a596c73
Requests from JOSS: Improved help and functions documentation, added a link to repository in JOSS paper, added CONTRIBUTING.md.
