+++ R/fct_full_factorial.R
@@ -30,9 +30,12 @@
 #'
 #' @importFrom stats runif na.omit
 #'
-#' @return A list with information two elements: \code{infoDesign} is a 
-#' and \code{fieldBook} is a data frame with the full factorial field book.
-#'
+#' @return A list with two elements.
+#' \itemize{
+#'   \item \code{infoDesign} is a list with information on the design parameters.
+#'   \item \code{fieldBook} is a data frame with the full factorial field book.
+#' }
+#' 
 #' @references
 #' Federer, W. T. (1955). Experimental Design. Theory and Application. New York, USA. The
 #' Macmillan Company.
