+++ R/fct_alpha_lattice.R
@@ -24,8 +24,11 @@
 #' @importFrom stats runif na.omit
 #' 
 #' 
-#' @return A list with information on the design parameters. 
-#' @return Data frame with the alpha design field book.
+#' @return A list with two elements.
+#' \itemize{
+#'   \item \code{infoDesign} is a list with information on the design parameters.
+#'   \item \code{fieldBook} is a data frame with the alpha design field book.
+#' }
 #'
 #'
 #' @references
