+++ R/fct_square_lattice.R
@@ -21,8 +21,11 @@
 #' 
 #' @importFrom stats runif na.omit
 #'
-#' @return A list with information on the design parameters.
-#' @return Data frame with the square lattice design field book.
+#' @return A list with two elements.
+#' \itemize{
+#'   \item \code{infoDesign} is a list with information on the design parameters.
+#'   \item \code{fieldBook} is a data frame with the square lattice design field book.
+#' }
 #'
 #' @references
 #' Edmondson., R. N. (2021). blocksdesign: Nested and crossed block designs for factorial and
