### 2021-04-27 fa0859d
Requests from JOSS: Improved help documentation.
