### 2021-04-27 DidierMurilloF/FielDHub#14 @DidierMurilloF
@dlebauer and @Prof-ThiagoOliveira

I got all the suggestions from you on this issue. I pushed all of the changes. Please, let me know if you have any questions! Please, let me know if we can go ahead and close this issue. Thanks.

Best,

Didier Murillo.
