### 2021-04-27 DidierMurilloF/FielDHub#12 @DidierMurilloF
Dear @dlebauer,

I edited the paper and added a new section `Availability and Community Guidelines.`  Please, let me know if you have any questions. Thank you!

Best,

Didier Murillo.
