+++ R/mod_RowCol.R
@@ -85,7 +85,37 @@ mod_RowCol_ui <- function(id){
 #' @noRd 
 mod_RowCol_server <- function(id){
   moduleServer( id, function(input, output, session){
+    
     ns <- session$ns
+    
+    entryListFormat_RCD <- data.frame(ENTRY = 1:9, 
+                                       NAME = c(paste("Genotype", LETTERS[1:9], sep = "")))
+    entriesInfoModal_RCD <- function() {
+      modalDialog(
+        title = div(tags$h3("Important message", style = "color: red;")),
+        h4("Please, follow the format shown in the following example. Make sure to upload a CSV file!"),
+        renderTable(entryListFormat_RCD,
+                    bordered = TRUE,
+                    align = 'c',
+                    striped = TRUE),
+        h4("Users can use any set of entry numbers."),
+        easyClose = FALSE
+      )
+    }
+    
+    toListen <- reactive({
+      list(input$owndataRCD)
+    })
+    
+    observeEvent(toListen(), {
+      if (input$owndataRCD == "Yes") {
+        showModal(
+          shinyjqui::jqui_draggable(
+            entriesInfoModal_RCD()
+          )
+        )
+      }
+    })
     
     getData.rcd <- reactive({
       req(input$file.RCD)

[…]
