+++ R/fct_latin_square.R
@@ -181,137 +181,11 @@ latin_square <- function(t = NULL, reps = 1, plotNumber = 101,  planter = "serpe
     numberofColumns =  length(unique(ls.output$COLUMN)),
     rep = reps,
     locationName =  locationNames,
-    seed =  seed
+    seed =  seed,
+    idDesign = 3
   )
   output <- list(infoDesign =  parameters, squares = lsd.reps,
                  plotSquares = plotSquares, fieldBook = latin_design)
   class(output) <- "FielDHub"
   return(invisible(output))
-}
-
-#-----------------------------------------------------------------------
-# Print
-#-----------------------------------------------------------------------
-#' @rdname print.latin_square
-#' @method print latin_square
-#' @title Print an \code{FielDHub} object
-#' @usage \method{print}{latin_square}(x, n, ...)
-#' @aliases print.latin_square
-#' @description Prints information about latin square design
-#' @return an object inheriting from class \code{FielDHub}
-#' @param x an object inheriting from class
-#' @param n a single integer. If positive or zero, size for the
-#'   resulting object: number of elements for a vector (including
-#'   lists), rows for a matrix or data frame or lines for a function. If
-#'   negative, all but the n last/first number of elements of x.
-#'
-#' @param ... further arguments passed to \code{\link{head}}.
-#' @author Thiago de Paula Oliveira,
-#'   \email{thiago.paula.oliveira@@alumni.usp.br}
-#' @importFrom utils head str
-#' @examples
-#' # Example 1: Generates a latin square design with 4 treatments and 2 reps.
-#' latinSq1 <- latin_square(t = 4,
-#'                          reps = 2,
-#'                          plotNumber = 101,
-#'                          planter = "cartesian",
-#'                          seed = 1980)
-#' print(latinSq1)
-#'
-#' @export
-
-print.latin_square <- function(x, n=10, ...){
-  #---------------------------------------------------------------------
-  # Parameters
-  #---------------------------------------------------------------------
-  cat("Latin square design:", "\n\n")
-  cat("Information on the design parameters:", "\n")
-  str(x$infoDesign)
-  #---------------------------------------------------------------------
-  # Head
-  #---------------------------------------------------------------------
-  nr <- nrow(x$fieldBook)
-  nhead <- min(n, nr)
-  if (n < 0) {
-    nhead_print <- nr + n
-  }else {
-    nhead_print <- nhead
-  }
-  cat("\n",  nhead_print,
-      "First observations of the data frame with the latin_square field book:",
-      "\n")
-  print(head(x$fieldBook, n=nhead, ...))
-}
-
-#-----------------------------------------------------------------------
-# Summary
-#-----------------------------------------------------------------------
-#' @rdname summary.latin_square
-#' @method summary latin_square
-#' @title Summary an \code{latin_square} object
-#' @usage \method{summary}{latin_square}(object, ...)
-#' @aliases summary.latin_square
-#' @description Summarise information on the design parameters, and data
-#'   frame structure
-#' @return an object inheriting from class \code{summary.latin_square}
-#' @param object an object inheriting from class
-#'   \code{FielDHub}
-#'
-#' @param ... Unused, for extensibility
-#' @author Thiago de Paula Oliveira,
-#'   \email{thiago.paula.oliveira@@alumni.usp.br}
-#'
-#' @examples
-#' # Example 1: Generates a latin square design with 4 treatments and 2 reps.
-#' latinSq1 <- latin_square(t = 4,
-#'                          reps = 2,
-#'                          plotNumber = 101,
-#'                          planter = "cartesian",
-#'                          seed = 1980)
-#' summary(latinSq1)
-#'
-#'
-#' @export
-summary.latin_square <- function(object, ...) {
[…]
