+++ R/mod_Square_Lattice.R
@@ -131,7 +131,7 @@ mod_Square_Lattice_server <- function(id){
                     bordered = TRUE,
                     align = 'c',
                     striped = TRUE),
-        h4("Users can use any set of entry numbers."),
+        h4("Entry numbers can be any set of consecutive positive numbers."),
         easyClose = FALSE
       )
     }

[…]
