+++ R/utils-pipe.R
@@ -1,6 +1,11 @@
 #' Pipe operator
 #'
 #' See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
+#' 
+#' @param lhs A value or the magrittr placeholder.
+#' @param rhs A function call using the magrittr semantics.
+#' 
+#' @return It depends on the arguments \code{lhs} and \code{rhs}.
 #'
 #' @name %>%
 #' @rdname pipe
