+++ R/run_app.R
@@ -1,12 +1,13 @@
 #' Run the Shiny Application
 #'
-#' @return A shiny app.
-#' 
+#' @return A shiny app object
+#' @param ... Unused, for extensibility
 #' 
 #' @export
 #' @importFrom shiny shinyApp
 #' @importFrom golem with_golem_options
 run_app <- function(
+  ...
 ) {
   with_golem_options(
     app = shinyApp(
