### 2021-05-12 #3122 @DidierMurilloF
Dear @csoneson,

Thank you very much for following up on the progress of our App. We have addressed all the suggestions that reviewers have made for the first release of FielDHub and we are committed to doing all the improvements suggested by reviewers for the next version. We are very pleased with all the improvements that have been suggested by the reviewers @Prof-ThiagoOliveira, @dlebauer and we believe this has made our application stronger and better.  We are looking forward to the acceptance of our App at the JOSS journal. Again, thank you very much and we look forward to hearing from you soon.

Best,

Didier Murillo.
