### 2021-05-13 #3122 @Prof-ThiagoOliveira
I agree with @dlebauer. You've done a great job and addressed all our comments. I've checked the package, and it is ready to be accepted as a publication on the JOSS. I am happy to see you have gone beyond our comments. Please, @csoneson, consider the paper accepted by my side.
