### 2021-05-15 #3122 @csoneson
@DidierMurilloF - this looks good! The only thing I would suggest is on line 27 in the paper, to change ``` it allows users to specify the dimensions of the field as row and columns``` to ``` it allows users to specify the dimensions of the field (the number of rows and columns)``` or something along those lines.

The next steps would then be as follows:
- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Archive the release in Zenodo or a similar service
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should match those of the paper. You may also add the authors' ORCID.
