### 2021-05-17 #3122 @DidierMurilloF
Dear @csoneson

I did the change you suggested in the paper, thank you. I am not familiar with the steps above, but I will learn about them as soon as possible.

Could you please indicate how should we tag the release of the software?

Best,

Didier Murillo
