### 2021-05-17 #3122 @csoneson
Sure! Here's how you create a release: https://docs.github.com/en/github/administering-a-repository/managing-releases-in-a-repository. We prefer that you tag it with the version of the package (e.g., `0.1.0`).

And here's a guide to archiving the repository in Zenodo: https://guides.github.com/activities/citable-code/
