+++ R/utils-pipe.R
@@ -5,7 +5,7 @@
 #' @param lhs A value or the magrittr placeholder.
 #' @param rhs A function call using the magrittr semantics.
 #' 
-#' @return It depends on the arguments \code{lhs} and \code{rhs}.
+#' @return No return value, called for side effects.
 #'
 #' @name %>%
 #' @rdname pipe
