### 2021-05-17 #3122 @DidierMurilloF
Dear @csoneson,

Thank you for your help. I got all steps listed above. Now `FielDHub` is available at Zenodo.

The corresponding DIO is the following,

http://doi.org/10.5281/zenodo.4768155

I also did small changes to the title of the paper to match with the release at R CRAN. Thank you!

Best,

Didier Murillo.
