### 2021-05-18 1be8a16
Update paper.bib
Some fixes to author names in JOSS paper bibliography file
