+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @Manual{agricolae_cit,
     title = {{agricolae}: Statistical Procedures for Agricultural Research},
-    author = {{Felipe de Mendiburu} and {Muhammad Yaseen}},
+    author = {Felipe {de Mendiburu} and Muhammad Yaseen},
     year = {2020},
     note = {R package version 1.4.0},
     url = {https://cran.r-project.org/package=agricolae},
@@ -9,7 +9,7 @@
 @Manual{blocksdesign_cit,
     title = {{blocksdesign}: Nested and Crossed Block Designs for Factorial and Unstructured
 Treatment Sets},
-    author = {R. N. Edmondson.},
+    author = {R. N. Edmondson},
     year = {2021},
     note = {R package version 4.8},
     url = {https://CRAN.R-project.org/package=blocksdesign},
@@ -37,14 +37,14 @@ Treatment Sets},
 }
 
 @book{Federer_cit,
-  	Author = {{Federer}, W.T.},
+  	Author = {Federer, W. T.},
   	Publisher = {The Macmillan Company.},
   	Title = {{Experimental Design. Theory and Application. New York, USA.}},
   	Year = 1955
 }
 
 @Article{Cullis2006_cit,
-    author = {Cullis, B.R., Smith, A.B. and Coombes, N.E.},
+    author = {B. R. Cullis and A. B. Smith and N. E. Coombes},
     title = {On the design of early generation variety trials with correlated data.},
     journal = {Journal of Agricultural, Biological, and Environmental Statistics},
     volume = {11},
