### 2021-05-18 #3122 @kyleniemeyer
Hi @DidierMurilloF, I'm the AEIC on duty this week doing some final checks before publishing. I noticed some minor issues in the references, could you merge this PR I made to fix them? https://github.com/DidierMurilloF/FielDHub/pull/16
